Add the proper conversions from 64 bit types to 32 bit types, or uses 64 bit types when no conversion to 32 bits is required.  This fixes compiler warnings for implicit shortening from 64 to 32 bits and fixes numerous bugs in implicit conversion to pointers.
Replace mkrf with extconf.rb and rake-compiler.  mkrf does not generate the correct link command OS X 10.7.

Replace version checks for feature checks.  No more -DRUBY_VERSION or -DWIN32.

Remove redefinition of NUM2* in ext/common/common.h.  The redefinition had the wrong type and Ruby 1.9 seems to use faster inline versions like the redefinitions were attempting to provide.

Switch to dlopen() from deprecated NSAddImage() and friends for OS X.

Fix bug in glut_TabletButtonFuncCallback that did not include the function to call

Fix bug in ARY2CMAT macro which wanted 5 arguments but only used and was called with 4

Switch to Gem::PackageTask from deprecated Rake::GemPackageTask

+++ ext/gl/gl-1.5.c
@@ -129,7 +129,7 @@ VALUE obj,arg1,arg2,arg3;
 	target = (GLenum)NUM2INT(arg1);
 	offset = (GLintptr)NUM2INT(arg2);
 	size = (GLsizeiptr)NUM2INT(arg3);
-	data = allocate_buffer_with_string(size);
+	data = allocate_buffer_with_string((long)size);
 	fptr_glGetBufferSubData(target,offset,size,(GLvoid *)RSTRING_PTR(data));
 	CHECK_GLERROR
 	return data;
